+++ src/components/ErrorNotification/index.tsx
@@ -1,0 +1,1 @@
+export * from './ErrorNotification';
