+++ src/components/TodoFooter/TodoFooter.tsx
@@ -1,0 +1,60 @@
+import React from 'react';
+import { Todo } from '../../types/Todo';
+
+type Props = {
+  todos: Todo[];
+  currentFilter: 'all' | 'active' | 'completed';
+  onFilterChange: (filter: 'all' | 'active' | 'completed') => void;
+};
+
+export const TodoFooter: React.FC<Props> = ({
+  todos,
+  currentFilter,
+  onFilterChange,
+}) => {
+  return (
+    <footer className="todoapp__footer" data-cy="Footer">
+      <span className="todo-count" data-cy="TodosCounter">
+        {todos.filter(todo => !todo.completed).length} items left
+      </span>
+
+      <nav className="filter" data-cy="Filter">
+        <a
+          href="#/"
+          className={`filter__link ${currentFilter === 'all' ? 'selected' : ''}`}
+          data-cy="FilterLinkAll"
+          onClick={() => onFilterChange('all')}
+        >
+          All
+        </a>
+
+        <a
+          href="#/active"
+          className={`filter__link ${currentFilter === 'active' ? 'selected' : ''}`}
+          data-cy="FilterLinkActive"
+          onClick={() => onFilterChange('active')}
+        >
+          Active
+        </a>
+
+        <a
+          href="#/completed"
+          className={`filter__link ${currentFilter === 'completed' ? 'selected' : ''}`}
+          data-cy="FilterLinkCompleted"
+          onClick={() => onFilterChange('completed')}
+        >
+          Completed
+        </a>
+      </nav>
+
+      <button
+        type="button"
+        disabled={todos.every(todo => !todo.completed)}
+        className="todoapp__clear-completed"
+        data-cy="ClearCompletedButton"
+      >
+        Clear completed
+      </button>
+    </footer>
+  );
+};
