+++ src/components/TodoFooter/index.tsx
@@ -1,0 +1,1 @@
+export * from './TodoFooter';
